# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.5",
    "title": "WebFM - Functional Mapping on the web",
    "contact": { "email": "croneecoglab@gmail.com" },
    "license": {
      "name": "GPL v3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "host": "zappa.neuro.jhu.edu:8090",
  "basePath": "/",
  "tags": [
    {
      "name": "brain",
      "description": "Everything about your Pets",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/api/brain/{subject}": {
      "get": {
        "tags": ["brain"],
        "summary": "Sends a 2D brain reconstruction to the browser",
        "description": "AS",
        "operationId": "getBrain",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "subject",
            "in": "path",
            "description": "ID of the subject to return",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": { "description": "Invalid ID supplied" },
          "404": { "description": "Pet not found" }
        }
      },
      "post": {
        "tags": ["brain"],
        "summary": "Add a new brain",
        "description": "",
        "operationId": "addBrain",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Add a 2D brain to the server",
            "required": true
          }
        ],
        "responses": { "405": { "description": "Invalid input" } }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about WebFM",
    "url": "https://github.com/cronelab/webfm"
  }
}